requirements.txt
# Core
Django==5.2.3
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
djoser==2.3.1

# Database
psycopg2-binary==2.9.10
mysqlclient==2.2.7

# CORS
django-cors-headers==3.7.0

# Filtering
django-filter==25.2

# Cloudinary
cloudinary==1.42.1
django-cloudinary-storage>=0.3.0

# Caching
django-redis==5.4.0

# Static files
whitenoise==6.11.0

# Config
python-decouple==3.8

# Server
gunicorn==23.0.0

# Sanitization
bleach==6.3.0
